Relocates files and exports
