--- index.js
@@ -1,21 +1,0 @@
-import AriaComponent from './src/AriaComponent';
-import Listbox from './src/Listbox';
-import MenuBar from './src/MenuBar';
-import Menu from './src/Menu';
-import Popup from './src/Popup';
-import Disclosure from './src/Disclosure';
-import Dialog from './src/Dialog';
-import Tablist from './src/Tablist';
-import MenuButton from './src/MenuButton';
-
-export {
-  AriaComponent,
-  Listbox,
-  MenuBar,
-  Menu,
-  Popup,
-  Disclosure,
-  Dialog,
-  Tablist,
-  MenuButton,
-};
