+++ src/Disclosure/Disclosure.modules.test.js
@@ -1,12 +1,13 @@
 /* eslint-disable max-len */
-import Disclosure from '.';
-import {
-  ComponentConnector,
-  ManageTabIndex,
-  UseButtonRole,
-  UseHiddenAttribute,
-} from './modules';
-import { events } from '../../.jest/events';
+import
+  Disclosure,
+  {
+    ComponentConnector,
+    ManageTabIndex,
+    UseButtonRole,
+    UseHiddenAttribute,
+  } from './';
+import { events } from '@/.jest/events';
 
 const {
   keydownEnter,
