+++ src/MenuButton/MenuButton.js
@@ -1,0 +1,8 @@
+/**
+ * Error: The MenuButton component is removed.
+ */
+export default class MenuButton {
+  constructor() {
+    throw new Error('[aria-components]: The MenuButton component is removed.');
+  }
+}
