+++ src/MenuButton/MenuButton.test.js
@@ -1,3 +1,3 @@
-import { MenuButton } from '../..';
+import MenuButton from '.';
 
 it('Should throw', () => expect(() => new MenuButton()).toThrow());
