+++ src/MenuButton/index.js
@@ -1,8 +1,3 @@
-/**
- * Error: The MenuButton component is removed.
- */
-export default class MenuButton {
-  constructor() {
-    throw new Error('[aria-components]: The MenuButton component is removed.');
-  }
-}
+import MenuButton from './MenuButton';
+
+export default MenuButton;
