+++ src/Tablist/Tablist.test.js
@@ -1,6 +1,6 @@
 /* eslint-disable max-len */
-import { Tablist } from '../..';
-import { events } from '../../.jest/events';
+import Tablist from '.';
+import { events } from '@/.jest/events';
 
 const {
   click,
